Add simple login lab and tests
The lab is currently barebones (HTML without any styling). This lab is
meant for practicing simple SQL injection. The login page is vulnerable
to SQL injection which allows for auth bypass. The profile page is
vulnerable to multi-statement SQL injection which allows for arbitrarily
modifying any table in the database. In particular, the grades for each
course can be modified.